requirements.txt
appnope==0.1.0
arrow==0.15.4
backports.shutil-get-terminal-size==1.0.0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==1.11.23
django-extensions==2.2.1
django-filter==2.0.0
django-oauth-toolkit==1.1.2
djangorestframework==3.9.4
enum34==1.1.6
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==19.9.0
httplib2==0.14.0
idna==2.8
ipython==5.8.0
ipython-genutils==0.2.0
Markdown==3.1.1
mysqlclient==1.4.6
oauth2client==4.1.3
oauthlib==3.1.0
pathlib2==2.3.4
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==1.0.16
psycopg2-binary==2.7.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scandir==1.10.0
simplegeneric==0.8.1
six==1.12.0
traitlets==4.3.2
typing==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.15.6
whitenoise==3.2.1
